life-in-weeks/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart